backend/requirements.txt
flask
flask-cors
tensorflow
pillow
opencv-python
numpy
gdown
